+++ src/pages/welcome-page/components/info-about-project/index.tsx
