+++ src/pages/welcome-page/components/info-about-project/info-about-project.css
