+++ src/pages/welcome-page/components/info-about-project/styled.ts
