+++ src/pages/welcome-page/welcome-page.tsx
@@ -1,5 +1,5 @@
 import './welcome-page.css';
-import { InfoAboutProject } from './components/infoAboutProject/index';
+import { InfoAboutProject } from './components/info-about-project/index';
 
 export const WelcomePage = () => (
   <div className="main-page">
